requirements.txt
openai>=0.28.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart
requests
aiohttp
email-validator>=2.2.0
PyJWT>=2.10.1
passlib>=1.7.4